[2.0.0-PRE-15] Hooked InteractiveChat properly & Improved code

+++ project/implementation-bukkit/src/main/kotlin/me/arasple/mc/trchat/module/internal/data/Database.kt
@@ -32,7 +32,7 @@ object Database {
         }
     }
 
-    @Schedule(delay = 100, period = 20 * 60 * 5, async = true)
+    @Schedule(delay = 100, period = (20 * 60 * 5).toLong(), async = true)
     @Awake(LifeCycle.DISABLE)
     fun save() {
         onlinePlayers().forEach { database.push(it.cast()) }
